+++ server/app.py
@@ -1,6 +1,5 @@
 import os
 from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy
 from TodosAPI import todos_api, Todo
 from sqlAlchemy_db_instance import db
 
@@ -20,28 +19,8 @@ def create_app():
     db.init_app(app)
     app.register_blueprint(todos_api)
 
-    @app.route('/')
-    def serve_vue_app():
-        """
-        Serve our vue app
-        """
-        return(render_template('index.html')) 
-
-    @app.after_request
-    def add_header(req):
-        """
-        Clear Cache for hot-reloading
-        """
-        req.headers["Cache-Control"] = "no-cache"
-        return req
-
     return app
 
-    def setup_database(app):
-        with app.app_context():
-            db.create_all()
-
-    if __name__ == '__main__':
-        app = create_app()
-        setup_database(app)
-        app.run()
+def setup_database(app):
+    with app.app_context():
+        db.create_all()
